fixed expon pbox constructor
