+++ src/pyuncertainnumber/pba/cbox.py
@@ -8,7 +8,8 @@ from .params import Params
 from .intervals import Interval
 import scipy
 from .cbox_Leslie import cbox_from_extredists, cbox_from_pseudosamples
-from .constructors import pbox_from_extredists, pbox_from_pseudosamples
+from .constructors import pbox_from_pseudosamples
+from .pbox_abc import pbox_from_extredists
 import pyuncertainnumber.pba.distributions as d
 
 if TYPE_CHECKING:
